admin-api/pyproject.toml
[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
httpx = "^0.26.0"
faker = "^22.6.0"
black = "^23.12.1"
ruff = "^0.1.11"

[tool.poetry]
name = "admin-api"
version = "0.1.0"
description = "ex-GPT Admin API - Management and monitoring system"
authors = ["Datastreams"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"
python-multipart = "^0.0.6"
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-settings = "^2.1.0"
httpx = "^0.26.0"
redis = "^5.0.1"
pandas = "^2.3.3"
openpyxl = "^3.1.5"
APScheduler = "^3.11.0"
cerbos = "^0.14.0"
qdrant-client = "^1.7.0"
minio = "^7.2.0"
PyPDF2 = "^3.0.0"
python-docx = "^1.1.0"
olefile = "^0.47"
psutil = "^7.1.1"
mlflow = "^2.10.0"
rq = "^2.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
